+++ src/views/BrowseView/BrowseView.test.js
@@ -4,23 +4,18 @@ import {
 } from '@folio/jest-config-stripes/testing-library/react';
 
 import { runAxeTest } from '@folio/stripes-testing';
-import {
-  useAuthoritiesBrowse,
-  useBrowseResultFocus,
-} from '@folio/stripes-authority-components';
+import { useAuthoritiesBrowse } from '@folio/stripes-authority-components';
 
 import BrowseView from './BrowseView';
 
 import Harness from '../../../test/jest/helpers/harness';
 import AuthoritiesLookup from '../../components/AuthoritiesLookup';
 
-const mockIsPaginationClicked = { current: false };
 const mockHandleLoadMore = jest.fn();
 
 jest.mock('@folio/stripes-authority-components', () => ({
   ...jest.requireActual('@folio/stripes-authority-components'),
   useAuthoritiesBrowse: jest.fn(),
-  useBrowseResultFocus: jest.fn(),
 }));
 
 jest.mock('../../components/AuthoritiesLookup', () => jest.fn(() => <div>AuthoritiesLookup</div>));
@@ -39,8 +34,6 @@ const renderBrowseView = (props = {}) => render(
 
 describe('Given BrowseView', () => {
   beforeEach(() => {
-    mockIsPaginationClicked.current = false;
-
     useAuthoritiesBrowse.mockReturnValue({
       authorities: [],
       hasNextPage: false,
@@ -50,10 +43,6 @@ describe('Given BrowseView', () => {
       handleLoadMore: mockHandleLoadMore,
       query: '(headingRef>="" or headingRef<"") and isTitleHeadingRef==false',
       totalRecords: 0,
-    });
-    useBrowseResultFocus.mockReturnValue({
-      resultsContainerRef: { current: null },
-      isPaginationClicked: mockIsPaginationClicked,
     });
   });
 
@@ -83,8 +72,8 @@ describe('Given BrowseView', () => {
       onNeedMoreData: expect.any(Function),
       onSubmitSearch: expect.any(Function),
       onLinkRecord: mockOnLinkRecord,
+      pagingOffset: 0,
       query: '(headingRef>="" or headingRef<"") and isTitleHeadingRef==false',
-      resultsContainerRef: { current: null },
       searchQuery: '',
       tenantId: '',
       totalRecords: 0,
@@ -95,14 +84,13 @@ describe('Given BrowseView', () => {
   });
 
   describe('when a user clicks on the pagination button', () => {
-    it('should invoke handleLoadMore and assign isPaginationClicked to true', async () => {
+    it('should invoke handleLoadMore', async () => {
       const args = [100, 95, 0, 'next'];
 
       renderBrowseView();
       await act(() => { AuthoritiesLookup.mock.calls[0][0].onNeedMoreData(...args); });
 
       expect(mockHandleLoadMore).toHaveBeenCalledWith(...args);
-      expect(mockIsPaginationClicked.current).toBeTruthy();
     });
   });
 });